+++ dwds/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.7.1-dev
+## 0.7.1
 
 - Fix a bug where we would try to create a new isolate even for a failed
   hot restart. This created a race condition that would lead to a crash.